+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
   allowing to extend an Array with zeros.
 - DMRG parameter ``'orthogonal_to'`` allows to calculate excited states for finite systems.
 - possibility to change the number of charges after creating LegCharges/Arrays
+- more general way to specify the order of sites in a :class:`tenpy.models.lattice.Lattice`.
 
 Changed
 ^^^^^^^